Moving view tracking down to SplitRecs component
