+++ src/app/containers/CpsRecommendations/SplitRecommendations/index.jsx
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { OptimizelyContext } from '@optimizely/react-sdk';
+import useViewTracker from '#hooks/useViewTracker';
 import { arrayOf, shape, number, string, func } from 'prop-types';
 import { storyItem } from '#models/propTypes/storyItem';
+import getEventTrackingData from '../RecommendationsPromoList/getEventTrackingData';
 import CpsRecommendations from '..';
 
 /*
@@ -8,7 +11,14 @@ import CpsRecommendations from '..';
  * This component should be removed after the experiment.
  */
 const SplitRecommendations = ({ items, ...props }) => {
-  const { part, splitRecsViewEventTracker } = props;
+  const { part } = props;
+  // OPTIMIZELY: 003_hindi_experiment_feature.
+  const { optimizely } = useContext(OptimizelyContext);
+  const eventTrackingData = getEventTrackingData();
+  const splitRecsViewEventTracker = useViewTracker({
+    ...eventTrackingData.block,
+    ...(optimizely && { optimizely }),
+  });
 
   if (!Array.isArray(items)) {
     return null;
@@ -27,13 +37,7 @@ const SplitRecommendations = ({ items, ...props }) => {
 
   if (part === 2) {
     const secondPart = items && items.slice(2, 4);
-    return (
-      <CpsRecommendations
-        {...props}
-        items={secondPart}
-        splitRecsViewEventTracker={splitRecsViewEventTracker}
-      />
-    );
+    return <CpsRecommendations {...props} items={secondPart} />;
   }
 
   return null;
